move planet data js file from public to data subdirectory
